FIXUP: Change from entry count to memory size
This changes cache eviction from a count of entries to the total amount
of memory used by the cache (as approximated by the memory used by ets
for the ?CACHE table). This behavior reverts to match the thresholds
used by the previous version as well as let users make more direct
choices on how large to make their design document caches.