Import [node] config settings into /_nodes DB
This patch allows administrators to configure settings in the [node]
block of the server configuration and have those settings propagated to
the internal nodes database. A current use case for this functionality
is to configure the `zone` of the node for use in cluster placement
topologies. Additional attributes can be written directly into the
document in the nodes DB, but in case of a conflict between the entry
in the database and the entry in the server configuration the latter
will be preferred (and the attribute in the database overwritten).

Related to #1338.