Optimize skip for partitioned queries
'skip' is implemented efficiently at the worker level but we've
disabled it for clustered views because of the multiple shards (and
not being able to calculate the right skip value to pass to each
worker). With a partitioned query, this problem is gone, as the value
the query specifies will be the right value for all workers (as they
hit the same shard range).

This commit removes the old fix_skip_and_limit function from
fabric_rpc and moves the logic up to the coordinators.