Move work from couch_server to open_async
This moves a lot of the mundane work out of couch_server's main loop and
into the open_async process to minimize the number of reductions and
messages required to handle individual messages arrivng in
couch_server's mailbox. The three main bits of work are: getting the
engine for the database, logging error results, and updating the open
time histogram.